requirements.txt
python-dotenv~=1.0.1
aiohttp~=3.11.11
base58~=2.1.1